# [WIP] Stop using reflection to deserialize JSON