--- a/src/client/Microsoft.Identity.Client/Cache/Items/MsalCredentialCacheItemBase.cs
+++ b/src/client/Microsoft.Identity.Client/Cache/Items/MsalCredentialCacheItemBase.cs
@@ -14,9 +14,9 @@ namespace Microsoft.Identity.Client.Cache.Items
 
         internal override void PopulateFieldsFromJObject(JObject j)
         {
-            CredentialType = JsonUtils.ExtractExistingOrEmptyString(j, StorageJsonKeys.CredentialType);
-            ClientId = JsonUtils.ExtractExistingOrEmptyString(j, StorageJsonKeys.ClientId);
-            Secret = JsonUtils.ExtractExistingOrEmptyString(j, StorageJsonKeys.Secret);
+            CredentialType = JsonHelper.ExtractExistingOrEmptyString(j, StorageJsonKeys.CredentialType);
+            ClientId = JsonHelper.ExtractExistingOrEmptyString(j, StorageJsonKeys.ClientId);
+            Secret = JsonHelper.ExtractExistingOrEmptyString(j, StorageJsonKeys.Secret);
 
             // Important: this MUST be last, since it will extract the AdditionalFieldsJson
             // after all other fields are read.