--- a/src/client/Microsoft.Identity.Client/Instance/Discovery/InstanceDiscoveryMetadataEntry.cs
+++ b/src/client/Microsoft.Identity.Client/Instance/Discovery/InstanceDiscoveryMetadataEntry.cs
@@ -1,8 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Linq;
 using Microsoft.Identity.Client.Utils;
 using Microsoft.Identity.Json;
+using Microsoft.Identity.Json.Linq;
 
 namespace Microsoft.Identity.Client.Instance.Discovery
 {
@@ -10,23 +12,38 @@ namespace Microsoft.Identity.Client.Instance.Discovery
     [Preserve(AllMembers = true)]
     internal sealed class InstanceDiscoveryMetadataEntry : IJsonSerializable<InstanceDiscoveryMetadataEntry>
     {
-        [JsonProperty(PropertyName = "preferred_network")]
+        private const string PreferredNetworkPropertyName = "preferred_network";
+        private const string PreferredCachePropertyName = "preferred_cache";
+        private const string AliasesPropertyName = "aliases";
+
+        [JsonProperty(PropertyName = PreferredNetworkPropertyName)]
         public string PreferredNetwork { get; set; }
 
-        [JsonProperty(PropertyName = "preferred_cache")]
+        [JsonProperty(PropertyName = PreferredCachePropertyName)]
         public string PreferredCache { get; set; }
 
-        [JsonProperty(PropertyName = "aliases")]
+        [JsonProperty(PropertyName = AliasesPropertyName)]
         public string[] Aliases { get; set; }
 
-        public InstanceDiscoveryMetadataEntry FromJsonString(string json)
+        public InstanceDiscoveryMetadataEntry DeserializeFromJson(string json)
         {
-            throw new System.NotImplementedException();
+            JObject jObject = JObject.Parse(json);
+
+            PreferredNetwork = jObject[PreferredNetworkPropertyName]?.ToString();
+            PreferredCache = jObject[PreferredCachePropertyName]?.ToString();
+            Aliases = ((JArray)jObject[AliasesPropertyName]).Select(c => (string)c).ToArray();
+
+            return this;
         }
 
-        public string ToJsonString(InstanceDiscoveryMetadataEntry objectToSerialize)
+        public string SerializeToJson()
         {
-            throw new System.NotImplementedException();
+            JObject jObject = new JObject(
+                new JProperty(PreferredNetworkPropertyName, PreferredNetwork),
+                new JProperty(PreferredCachePropertyName, PreferredCache),
+                new JProperty(AliasesPropertyName, new JArray(Aliases)));
+
+            return jObject.ToString(Formatting.None);
         }
     }
 }