--- a/src/client/Microsoft.Identity.Client/Instance/Discovery/InstanceDiscoveryResponse.cs
+++ b/src/client/Microsoft.Identity.Client/Instance/Discovery/InstanceDiscoveryResponse.cs
@@ -1,9 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Linq;
 using Microsoft.Identity.Client.OAuth2;
-using Microsoft.Identity.Json;
 using Microsoft.Identity.Client.Utils;
+using Microsoft.Identity.Json;
+using Microsoft.Identity.Json.Linq;
 
 namespace Microsoft.Identity.Client.Instance.Discovery
 {
@@ -11,20 +13,34 @@ namespace Microsoft.Identity.Client.Instance.Discovery
     [Preserve(AllMembers = true)]
     internal sealed class InstanceDiscoveryResponse : OAuth2ResponseBase, IJsonSerializable<InstanceDiscoveryResponse>
     {
-        [JsonProperty(PropertyName = "tenant_discovery_endpoint")]
+        private const string TenantDiscoveryEndpointPropertyName = "tenant_discovery_endpoint";
+        private const string MetadataPropertyName = "metadata";
+
+        [JsonProperty(PropertyName = TenantDiscoveryEndpointPropertyName)]
         public string TenantDiscoveryEndpoint { get; set; }
 
-        [JsonProperty(PropertyName = "metadata")]
+        [JsonProperty(PropertyName = MetadataPropertyName)]
         public InstanceDiscoveryMetadataEntry[] Metadata { get; set; }
 
-        public InstanceDiscoveryResponse FromJsonString(string json)
+        public new InstanceDiscoveryResponse DeserializeFromJson(string json)
         {
-            throw new System.NotImplementedException();
+            JObject jObject = JObject.Parse(json);
+
+            TenantDiscoveryEndpoint = jObject[TenantDiscoveryEndpointPropertyName]?.ToString();
+            Metadata = ((JArray)jObject[MetadataPropertyName]).Select(c => new InstanceDiscoveryMetadataEntry().DeserializeFromJson(c.ToString())).ToArray();
+            base.DeserializeFromJson(json);
+
+            return this;
         }
 
-        public string ToJsonString(InstanceDiscoveryResponse objectToSerialize)
+        public new string SerializeToJson()
         {
-            throw new System.NotImplementedException();
+            JObject jObject = new JObject(
+                new JProperty(TenantDiscoveryEndpointPropertyName, TenantDiscoveryEndpoint),
+                new JProperty(MetadataPropertyName, new JArray(Metadata.Select(i => JObject.Parse(i.SerializeToJson())))),
+                JObject.Parse(base.SerializeToJson()).Properties());
+            
+            return jObject.ToString(Formatting.None);
         }
     }
 }