--- a/src/client/Microsoft.Identity.Client/Instance/Discovery/InstanceDiscoveryResponse.cs
+++ b/src/client/Microsoft.Identity.Client/Instance/Discovery/InstanceDiscoveryResponse.cs
@@ -3,17 +3,28 @@
 
 using Microsoft.Identity.Client.OAuth2;
 using Microsoft.Identity.Json;
+using Microsoft.Identity.Client.Utils;
 
 namespace Microsoft.Identity.Client.Instance.Discovery
 {
     [JsonObject]
     [Preserve(AllMembers = true)]
-    internal sealed class InstanceDiscoveryResponse : OAuth2ResponseBase
+    internal sealed class InstanceDiscoveryResponse : OAuth2ResponseBase, IJsonSerializable<InstanceDiscoveryResponse>
     {
         [JsonProperty(PropertyName = "tenant_discovery_endpoint")]
         public string TenantDiscoveryEndpoint { get; set; }
 
         [JsonProperty(PropertyName = "metadata")]
         public InstanceDiscoveryMetadataEntry[] Metadata { get; set; }
+
+        public InstanceDiscoveryResponse FromJsonString(string json)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public string ToJsonString(InstanceDiscoveryResponse objectToSerialize)
+        {
+            throw new System.NotImplementedException();
+        }
     }
 }