… src/client/Microsoft.Identity.Client/Internal/ClientInfo.cs
… src/client/Microsoft.Identity.Client/Internal/IdToken.cs
… src/client/Microsoft.Identity.Client/MsalServiceException.cs
… src/client/Microsoft.Identity.Client/OAuth2/MsalTokenResponse.cs
… src/client/Microsoft.Identity.Client/OAuth2/OAuth2Client.cs
… src/client/Microsoft.Identity.Client/OAuth2/OAuth2ResponseBase.cs
… src/client/Microsoft.Identity.Client/Region/LocalImdsErrorResponse.cs
… src/client/Microsoft.Identity.Client/Region/RegionDiscoveryProvider.cs
… src/client/Microsoft.Identity.Client/Utils/IJsonSerializable.cs
… src/client/Microsoft.Identity.Client/Utils/JsonHelper.cs
… src/client/Microsoft.Identity.Client/Utils/JsonUtils.cs
… tests/Microsoft.Identity.Test.Integration.netfx/HeadlessTests/AuthorityTests.cs
… tests/Microsoft.Identity.Test.Performance/JsonTests.cs
… tests/Microsoft.Identity.Test.Performance/Microsoft.Identity.Test.Performance.csproj
… tests/Microsoft.Identity.Test.Performance/Program.cs
… tests/Microsoft.Identity.Test.Unit/CacheTests/TokenCacheTests.cs
… tests/Microsoft.Identity.Test.Unit/CacheTests/UnifiedSchemaValidationTests.cs
… tests/Microsoft.Identity.Test.Unit/CoreTests/InstanceTests/InstanceProviderTests.cs
… tests/Microsoft.Identity.Test.Unit/UtilTests/JsonHelperTests.cs